+++ addons/crm/static/src/core/channel_commands.js
