+++ addons/crm/static/tests/composer_tests.js
