+++ addons/crm_livechat/__manifest__.py
@@ -19,9 +19,6 @@
     'auto_install': True,
     'license': 'LGPL-3',
     'assets': {
-        'web.assets_backend': {
-            'crm_livechat/static/src/core/*',
-        },
         'web.qunit_suite_tests': [
             'crm_livechat/static/tests/**/*.js',
         ],
